% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amznDailyPrices2020}
\alias{amznDailyPrices2020}
\title{Daily adjusted closing price for Amazon stock}
\format{
A \samp{xts} object:
\describe{
\item{AMZN}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=AMZN+Historical+Prices}
}
\usage{
amznDailyPrices2020
}
\description{
A \samp{xts} object of daily adjusted closing price for Amazon from 1998-01-02
to 2020-12-30.
}
\keyword{datasets}
