% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amznMonthlyPrices}
\alias{amznMonthlyPrices}
\title{Monthly adjusted closing price for Amazon stock}
\format{A \samp{xts} object on Jan-1998 to Jan-2015 containing:
\describe{
  \item{AMZN}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=AMZN+Historical+Prices}
}
\usage{
amznMonthlyPrices
}
\description{
A \samp{xts} object of monthly adjusted closing price for Amazon from Jan-1998 to Jan-2015.
}
\keyword{datasets}

