% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baDailyPrices2020}
\alias{baDailyPrices2020}
\title{Daily adjusted closing price for Boeing stock}
\format{
A \samp{xts} object:
\describe{
\item{BA}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=BA+Historical+Prices}
}
\usage{
baDailyPrices2020
}
\description{
A \samp{xts} object of daily adjusted closing price for Boeing from 1993-01-04 to 2020-12-30.
}
\keyword{datasets}
