% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gwn_model.R
\name{coef.gwn_model}
\alias{coef.gwn_model}
\title{Extract Model Coefficients from GWN Model}
\usage{
\method{coef}{gwn_model}(
  object,
  which_coef = c("mean", "volatility", "variance", "covariance", "correlation"),
  ...
)
}
\arguments{
\item{object}{of class gwn_model}

\item{which_coef}{character, one of "mean", "volatility", "variance", "covariance",
or "correlation"}

\item{...}{additional arguments passed to the function}
}
\value{
Coefficients extracted from the model \code{object}
}
\description{
Extract the specified coefficients from a GWN model.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
gwn_fit = estimate_gwn_model(exampleRetS[,"AMZN"])
gwn_fit2 = estimate_gwn_model(exampleRetS)
coef(gwn_fit, which_coef="mean")
}
\author{
Eric Zivot
}
