% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{costDailyPrices}
\alias{costDailyPrices}
\title{Daily adjusted closing price for Costco stock}
\format{
A \samp{xts} object containing:
\describe{
\item{COST}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=COST+Historical+Prices}
}
\usage{
costDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Costco from 1993-01-04 to 2014-12-31.
}
\keyword{datasets}
