% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{efficient_frontier}
\alias{efficient_frontier}
\title{Compute efficient frontier of risky assets}
\usage{
efficient_frontier(
  er_vec,
  cov_mat,
  nport = 20,
  alpha_min = -0.5,
  alpha_max = 1.5,
  shorts = TRUE
)
}
\arguments{
\item{er_vec}{\samp{N x 1} vector of expected returns}

\item{cov_mat}{\samp{N x N} return covariance matrix}

\item{nport}{scalar, number of efficient portfolios to compute}

\item{alpha_min}{minimum value of \samp{alpha}, default is \samp{-.5}}

\item{alpha_max}{maximum value of \samp{alpha}, default is \samp{1.5}}

\item{shorts}{logical, if \samp{TRUE} then short sales (negative portfolio weights)
are allowed. If \samp{FALSE} then no asset is allowed to be sold short}
}
\value{
An object of class \samp{Markowitz} with the following components:
\itemize{
\item \code{call} captures function call
\item \code{er_vec} \verb{nport x 1} vector of expected returns of efficient porfolios
\item \code{sd_vec} \verb{nport x 1} vector of standard deviations of efficient portfolios
\item \code{weight_mat} \verb{nport x N} matrix of weights of efficient portfolios
}
}
\description{
The function constructs the set of mean-variance efficient portfolios that either allow all
assets to be sold short or not allow any asset to be sold short. The returned object is of class
\samp{Markowitz} for which there are \samp{print}, \samp{summary} and \samp{plot} methods.
}
\details{
If short sales are allowed (negative weights) then the set of efficient portfolios of risky
assets can be computed as a convex combination of any two efficient portfolios. It is convenient
to use the global minimum variance portfolio as one portfolio and an efficient portfolio with
target expected return equal to the maximum expected return of the assets under consideration as
the other portfolio. Call these portfolios \eqn{m} and \eqn{x}, respectively. Then for any number
\samp{alpha}, another efficient portfolio can be computed as \eqn{z=\alpha m+(1-\alpha)x}. If
short sales are not allowed, then the set of efficient portfolios is computed by repeated calls
to the function \samp{efficient_portfolio()}, with \samp{shorts=FALSE}, for a grid of target
expected returns starting at the expected return of the global minimum variance portfolio (not
allowing short sales) and ending at the expected return equal to the maximum expected return of
the assets under consideration.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# tangency portfolio
tan_port <- tangency_portfolio(er, covmat, r_free)
# compute global minimum variance portfolio
gmin_port = globalmin_portfolio(er, covmat)

# compute portfolio frontier
ef <- efficient_frontier(er, covmat, alpha_min=-2,
                         alpha_max=1.5, nport=20)
attributes(ef)

plot(ef)
plot(ef, plot.assets=TRUE, col="blue", pch=16)
points(gmin_port$sd_port, gmin_port$er_port, col="green", pch=16, cex=2)
points(tan_port$sd_port, tan_port$er_port, col="red", pch=16, cex=2)
text(gmin_port$sd_port, gmin_port$er_port, labels="GLOBAL MIN", pos=2)
text(tan_port$sd_port, tan_port$er_port, labels="TANGENCY", pos=2)
sr_tan = (tan_port$er_port - r_free)/tan_port$sd_port
abline(a=r_free, b=sr_tan, col="green", lwd=2)

}
\author{
Eric Zivot
}
