% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_portfolio.R
\name{efficient_portfolio}
\alias{efficient_portfolio}
\title{Compute minimum variance portfolio subject to target return}
\usage{
efficient_portfolio(er_vec, cov_mat, target_return, shorts = TRUE)
}
\arguments{
\item{er_vec}{\eqn{N \times 1} vector of expected returns}

\item{cov_mat}{\eqn{N \times N} return covariance matrix}

\item{target_return}{scalar, target expected return}

\item{shorts}{logical, if \code{TRUE} then short sales (negative portfolio weights)
are allowed. If \code{FALSE} then no asset is allowed to be sold short.}
}
\value{
A \code{portfolio} object with the following components:
\itemize{
\item \code{call} captures function call
\item \code{er_port} portfolio expected return \eqn{\mu_p = \mathbf{x}'\boldsymbol{\mu}}
\item \code{sd_port} portfolio standard deviation \eqn{\sigma_p = \sqrt{\mathbf{x}'\boldsymbol{\Sigma}\mathbf{x}}}
\item \code{weight_vec} \eqn{N \times 1} vector of portfolio weights \eqn{\mathbf{x}}
\item \code{er_vec} \eqn{N \times 1} vector of expected returns \eqn{\boldsymbol{\mu}}
\item \code{cov_mat} \eqn{N \times N} return covariance matrix \eqn{\boldsymbol{\Sigma}}
}
}
\description{
Compute minimum variance portfolio subject to target return either allowing all assets to be sold
short or not allowing any asset to be sold short. The returned object is of class
\code{portfolio}.
}
\details{
A mean-variance efficient portfolio \eqn{\mathbf{x}} allowing short sales (negative weights) that achieves
a target expected return \eqn{\mu_0} solves the optimization problem:
\deqn{\min_{\mathbf{x}} \mathbf{x}'\boldsymbol{\Sigma} \mathbf{x} ~ s.t.~
\mathbf{x}'\mathbf{1}=1 ~ \mathrm{and} ~  \mathbf{x}'\boldsymbol{\mu}=\mu_0,}
for which there is an analytic solution using matrix
algebra. If short sales are not allowed then the portfolio is computed numerically using the
function \code{solve.QP()} from the \code{quadprog} package.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# compute efficient portfolio subject to target return
target_return = er["MSFT"]
e_port_msft = efficient_portfolio(er, covmat, target_return)
e_port_msft
summary(e_port_msft, risk_free=r_free)
plot(e_port_msft, col="blue")

# compute efficient portfolio subject to target return with no short sales
target_return = er["MSFT"]
e_port_msft_ns = efficient_portfolio(er, covmat, target_return, shorts=FALSE)
e_port_msft_ns
summary(e_port_msft_ns, risk_free=r_free)
plot(e_port_msft_ns, col="blue")

}
\author{
Eric Zivot
}
