% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gwn_model.R
\name{estimate_gwn_model}
\alias{estimate_gwn_model}
\title{Estimate Gaussian White Noise Model}
\usage{
estimate_gwn_model(x)
}
\arguments{
\item{x}{\verb{n.obs x n.vars} xts time series, matrix or vector of data. If a vector is provided,
it will be converted to a matrix with one column.`}
}
\value{
An object of class \code{gwn_model} with the following components:
\itemize{
\item \code{call} The matched function call
\item \code{muhat} Estimated mean vector \eqn{\hat{\boldsymbol{\mu}}}
\item \code{se_muhat} Estimated standard errors of the mean vector \eqn{\hat{\mathrm{se}}(\hat{\mu}_i),~ i=1,\ldots,n}
\item \code{Sigmahat} Estimated covariance matrix \eqn{\hat{\boldsymbol{\Sigma}}}
\item \code{Rhat} Estimated correlation matrix \eqn{\hat{\mathbf{R}}}
\item \code{sigma2hat} Estimated variances \eqn{\hat{\sigma}^2_i,~ i=1,\ldots,n}
\item \code{se_sigma2hat} Estimated standard errors of the sample variances \eqn{\hat{\mathrm{se}}(\hat{\sigma}^2_i),~ i=1,\ldots,n}
\item \code{sigmahat} Estimated standard deviations \eqn{\hat{\sigma}_i,~ i=1,\ldots,n}
\item \code{se_sigmahat} Estimated standard errors of the sample standard deviations \eqn{\hat{\mathrm{se}}(\hat{\sigma}_i),~ i=1,\ldots,n}
\item \code{covhat} Estimated covariances \eqn{\hat{\sigma}_{ij},~ i,j=1,\ldots,n}
\item \code{se_cov} Estimated standard errors of the sample covariances \eqn{\hat{\mathrm{se}}(\hat{\sigma}_{ij}),~ i,j=1,\ldots,n}
\item \code{rhohat} Estimated correlations \eqn{\hat{\rho}_{ij},~ i,j=1,\ldots,n}
\item \code{se_rhohat} Estimated standard errors of the sample correlations \eqn{\hat{\mathrm{se}}(\hat{\rho}_{ij}),~ i,j=1,\ldots,n}
\item \code{n.obs} Number of observations \eqn{n}
\item \code{ehat} Estimated residuals \eqn{\hat{\mathbf{e}}_t = \mathbf{R}_t - \hat{\boldsymbol{\mu}}}
}
}
\description{
Estimate Gaussian White Noise (GWN) model parameters from a vector or matrix of data.
The function estimates the mean vector, covariance matrix, and correlation matrix of the data,
as well as their standard errors. The function returns an object of class \code{gwn_model} which is a
list containing the estimated parameters and their standard errors, along with the number of observations.
There are corresponding \code{print}, \code{summary}, and \code{plot} methods for objects of class \code{gwn_model}.
}
\details{
The GWN model assumes that the returns are independent and identically distributed (i.i.d.) with a
multivariate normal distribution:
\deqn{\mathbf{R}_t \sim N(\boldsymbol{\mu}, \boldsymbol{\Sigma}),}
where \eqn{\mathbf{R}_t} is the \eqn{n \times 1} vector of returns at time \eqn{t}, \eqn{\boldsymbol{\mu}} is the
\eqn{n \times 1} mean vector, and \eqn{\boldsymbol{\Sigma}} is the \eqn{n \times n} covariance matrix.
The estimators of \eqn{\boldsymbol{\mu}} and \eqn{\boldsymbol{\Sigma}} are the sample mean and sample
covariance matrix, respectively.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
gwn_fit = estimate_gwn_model(exampleRetS[,"AMZN"])
gwn_fit2 = estimate_gwn_model(exampleRetS)
}
\references{
See Zivot, E. (2026) \emph{Introduction to Computational Finance and Financial Econometrics with R},
chapter 7 for a discussion of the Gaussian White Noise model and its estimation.
}
