% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sr.R
\name{estimate_sr}
\alias{estimate_sr}
\title{Estimate Sharpe Ratio}
\usage{
estimate_sr(x, r_f = 0)
}
\arguments{
\item{x, }{a vector or matrix of returns. If a matrix, each column is
treated as a separate asset.}

\item{r_f, }{scalar risk-free rate. The default is 0.}
}
\value{
A SRhat object with the following components:
\itemize{
\item \code{call} The function call
\item \code{date_range} The date range of the data
\item \code{srhat} The estimated Sharpe ratio \eqn{\widehat{\mathrm{SR}}}
\item \code{se_srhat} The estimated standard error of the Sharpe ratio \eqn{\hat{\mathrm{se}}(\widehat{\mathrm{SR}})}
\item \code{r_f} The risk-free rate \eqn{r_f}
\item \code{n.obs} The number of observations \eqn{n}
}
}
\description{
This function estimates the Sharpe ratio given a vector or matrix of returns.
The Sharpe ratio is estimated as the ratio of the mean excess return to the
standard deviation of excess returns. The function also computes the standard
error of the Sharpe ratio estimate using the delta method.
}
\details{
The estimated Sharpe ratio of an asset is defined as
\deqn{\widehat{\mathrm{SR}} = \frac{\hat{\mu} - r_f}{\hat{\sigma}}}
where \eqn{\hat{\mu}} is the estimated mean return, \eqn{r_f} is the risk-free rate, and
\eqn{\hat{\sigma}} is the estimated standard deviation of the returns. The estimated
standard error of the Sharpe ratio is computed using the delta method assuming that
returns are described by the gaussian white noise model.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
sr_hat = estimate_sr(exampleRetS)
names(sr_hat)
}
