% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_var.R
\name{estimate_var}
\alias{estimate_var}
\title{Estimate value at risk (VaR).}
\usage{
estimate_var(
  x,
  alpha = 0.05,
  w = 1e+05,
  var_method = c("normal", "empirical"),
  var_type = c("return", "dollar"),
  return_type = c("simple", "cc")
)
}
\arguments{
\item{x, }{a vector or matrix of returns. If a matrix, each column is
treated as a separate asset.}

\item{alpha, }{scalar tail probability \eqn{\alpha}.}

\item{w, }{scalar initial investment value.}

\item{var_method, }{character string specifying the method to use for estimating VaR.
Options are "normal" or "empirical".}

\item{var_type, }{character string specifying the type of VaR to estimate. Options ar
"dollar" or "return". If var_type is "dollar", then the VaR is estimated in dollar terms
by multiplying the estimated return VaR by the initial investment value. If var_type is
"return", then the VaR is estimated as the percentage loss in return terms.}

\item{return_type, }{character string specifying the type of returns. Options are
"simple" or "cc" (continuously compounded). If return_type is "cc", then the
log-normal VaR formula is used for normal VaR estimation.}
}
\value{
A VaRhat object with the following components:
\itemize{
\item \code{call} The function call
\item \code{date_range} The date range of the data
\item \code{alpha} The probability level
\item \code{w} The initial investment value
\item \code{var_method} The method used for estimating VaR
\item \code{var_type} The type of VaR estimated
\item \code{return_type} The type of returns used for estimating VaR
\item \code{VaRhat} The estimated value at risk \eqn{\widehat{\mathrm{VaR}}_{\alpha}}
\item \code{se_VaRhat} The estimated VaR standard error \eqn{\widehat{\mathrm{se}}(\widehat{\mathrm{VaR}}_{\alpha})}
}
}
\description{
This function estimates VaR assuming either a normal distribution,
log-normal distribution (i.e, normal distribution for continuously compounded returns),
or an empirical distribution. The VaR is estimated for a given probability level
and initial investment value. VaR is interpreted as the dollar loss of an initial investment
over a specified time period with a given probability level. For example, the 5\% VaR of a $100,000
investment over a one-month period is interpreted as "there is a 5\% chance that the portfolio will lose
VaR \emph{or more} over the next month."
}
\details{
Normal VaR for simple returns is estimated using the formula the left-tail \eqn{\alpha}-quantile
of a normal distribution:
\deqn{\widehat{\mathrm{VaR}}_{\alpha} = -(\hat{\mu} + \hat{\sigma} \times z_{\alpha})\times w,}
where \eqn{\hat{\mu}} and \eqn{\hat{\sigma}} are the sample mean and sample standard
deviation of the returns, \eqn{z_\alpha} is the \eqn{\alpha}-quantile of the
standard normal distribution, and \eqn{w} is the initial investment value.
Log-normal VaR (normal VaR for cc returns) is estimated as
\deqn{\widehat{\mathrm{VaR}}_{\alpha} = -(e^{\hat{\mu} + \hat{\sigma} z_{\alpha}} - 1) \times w.}
Empirical VaR is estimated as the negative of the empirical quantile of the returns
multiplied by the initial investment value:
\deqn{\widehat{\mathrm{VaR}}_{\alpha} = -\hat{q}_{\alpha} \times w,}
where \eqn{\hat{q}_{\alpha}} is the empirical \eqn{\alpha}-quantile of the returns.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
exampleRetC = log(1 + exampleRetS)
var_hat = estimate_var(exampleRetS, alpha=0.05, w=100000,
                      var_method="normal", var_type ="return",
                      return_type="simple")
names(var_hat)
}
