% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourpanelplot.R
\name{four_panel_plot}
\alias{four_panel_plot}
\title{Create four-panel diagnostic plot of returns}
\usage{
four_panel_plot(ret)
}
\arguments{
\item{ret}{Single data object (xts, zoo, matrix, data.frame) of returns. It is assumed that the
column has a name}
}
\description{
Four-panel plot showing histogram with normal curve overlaid, boxplot, sample ACF, and normal
qq-plot. Used for accessing the appropriateness of the Gaussian White Noise model for
returns. Inspired by the four-panel plot in \emph{Statistical Analysis of Financial Data with R} by
Rene Carmona.
}
\examples{
examplePrices=amznDailyPrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
four_panel_plot(exampleRetS)
}
