% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fourPanelPlot.R
\name{fourPanelPlot}
\alias{fourPanelPlot}
\title{Create four-panel plot of returns}
\usage{
fourPanelPlot(ret)
}
\arguments{
\item{ret}{Single data object (xts, zoo, matrix, data.frame) of returns. It is assumed that the
column has a name}
}
\description{
Four-panel plot showing histogram with normal curve overlaid, boxplot, sample ACF, and normal
qq-plot. Inspired by the four-panel plot in Statistical Analysis of Financial Data with R by
Rene Carmona.
}
\author{
Eric Zivot
}

