% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio.R
\name{get_portfolio}
\alias{get_portfolio}
\title{Create portfolio object}
\usage{
get_portfolio(er_vec, cov_mat, weight_vec)
}
\arguments{
\item{er_vec}{\eqn{N \times 1} vector of expected returns \eqn{\boldsymbol{\mu}}}

\item{cov_mat}{\eqn{N \times N} return covariance matrix \eqn{\boldsymbol{\Sigma}}}

\item{weight_vec}{\eqn{N \times 1} vector of portfolio weights \eqn{\mathbf{x}}}
}
\value{
An object of class \code{portfolio} with the following components:`
\itemize{
\item \code{call} captures function call
\item \code{er_port} portfolio expected return \eqn{\mu_p = \mathbf{x}'\boldsymbol{\mu}}
\item \code{sd_port} portfolio standard deviation \eqn{\sigma_p = \sqrt{\mathbf{x}'\boldsymbol{\Sigma}\mathbf{x}}}
\item \code{weight_vec} \eqn{N \times 1} vector of portfolio weights \eqn{\mathbf{x}}
\item \code{er_vec} \eqn{N \times 1} vector of expected returns \eqn{\boldsymbol{\mu}}
\item \code{cov_mat} \eqn{N \times N} return covariance matrix \eqn{\boldsymbol{\Sigma}}
}
}
\description{
Create a portfolio object from expected return vector \eqn{\boldsymbol{\mu}},
covariance matrix \eqn{\boldsymbol{\Sigma}}, and weight vector \eqn{\mathbf{x}}.
}
\details{
To specify a portfolio, an expected return vector and covariance matrix for the assets under
consideration as well as a vector of portfolio weights are needed. The result of
\code{get_portfolio} is a \code{portfolio} object, which is list with components for the portfolio
expected return, portfolio standard deviation, and portfolio weights. There are \code{print},
\code{summary} and \code{plot} methods.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# compute equally weighted portfolio
ew = rep(1,3)/3
equal_weight_portfolio = get_portfolio(er_vec=er,cov_mat=covmat,weight_vec=ew)
class(equal_weight_portfolio)
names(equal_weight_portfolio)
equal_weight_portfolio
summary(equal_weight_portfolio)
plot(equal_weight_portfolio, col="blue")

}
\author{
Eric Zivot
}
