% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPortfolio.R
\name{getPortfolio}
\alias{getPortfolio}
\title{Create portfolio object}
\usage{
getPortfolio(er, cov.mat, weights)
}
\arguments{
\item{er}{\samp{N x 1} vector of expected returns}

\item{cov.mat}{\samp{N x N} return covariance matrix}

\item{weights}{\samp{N x 1} vector of portfolio weights}
}
\value{
\item{call}{captures function call}
 \item{er}{portfolio expected return}
 \item{sd}{portfolio standard deviation}
 \item{weights}{\samp{N x 1} vector of portfolio weights}
}
\description{
Create a portfolio object from expected return vector, covariance matrix, and weight vector.
}
\details{
To specify a portfolio, an expected return vector and covariance matrix for the assets under
consideration as well as a vector of portfolio weights are needed. The result of \code{getPortfolio}
is a \samp{portfolio} object, which is list with components for the portfolio expected return,
portfolio standard deviation, and portfolio weights. There are \code{print}, \code{summary} and \code{plot}
methods.
}
\examples{
# construct the data
asset.names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset.names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset.names, asset.names)
er
covmat
r.free

# compute equally weighted portfolio
ew = rep(1,3)/3
equalWeight.portfolio = getPortfolio(er=er,cov.mat=covmat,weights=ew)
class(equalWeight.portfolio)
names(equalWeight.portfolio)
equalWeight.portfolio
summary(equalWeight.portfolio)
plot(equalWeight.portfolio, col="blue")
}
\author{
Eric Zivot
}

