% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalmin_portfolio.R
\name{globalmin_portfolio}
\alias{globalmin_portfolio}
\title{Compute global minimum variance portfolio}
\usage{
globalmin_portfolio(er_vec, cov_mat, shorts = TRUE)
}
\arguments{
\item{er_vec}{\eqn{N \times 1} vector of expected returns \eqn{\boldsymbol{\mu}}}

\item{cov_mat}{\eqn{N \times N} return covariance matrix \eqn{\boldsymbol{\Sigma}}}

\item{shorts}{logical, if \code{TRUE} then short sales (negative portfolio weights)
are allowed. If \code{FALSE} then no asset is allowed to be sold short.}
}
\value{
A \code{portfolio} object with the following components:
\itemize{
\item \code{call} captures function call
\item \code{er_port} portfolio expected return \eqn{\mu_m = \mathbf{m}'\boldsymbol{\mu}}`
\item \code{sd_port} portfolio standard deviation \eqn{\sigma_m = \sqrt{\mathbf{m}'\boldsymbol{\Sigma}\mathbf{m}}}
\item \code{weight_vec} \eqn{N \times 1} vector of portfolio weights \eqn{\mathbf{m}}
\item \code{er_vec} \eqn{N \times 1} vector of expected returns \eqn{\boldsymbol{\mu}}
\item \code{cov_mat} \eqn{N \times N} return covariance matrix \eqn{\boldsymbol{\Sigma}}
}
}
\description{
Compute global minimum variance portfolio given expected return vector \eqn{\boldsymbol{\mu}}
and covariance matrix \eqn{\boldsymbol{\Sigma}}. The resulting
portfolio \eqn{\mathbf{m}} can allow all assets to be shorted or not allow any assets to be shorted.
The returned  object is of class \code{portfolio}.
}
\details{
The global minimum variance portfolio \eqn{\mathbf{m}} allowing for short sales solves the optimization
problem:
\deqn{\min_{\mathbf{m}} \mathbf{m}'\boldsymbol{\Sigma} \mathbf{m} ~ s.t.~ \mathbf{m}'\mathbf{1}=1,}
for which there is an analytic solution using matrix algebra:
\deqn{\mathbf{m} = \frac{\boldsymbol{\Sigma}^{-1}\mathbf{1}}{\mathbf{1}'\boldsymbol{\Sigma}^{-1}\mathbf{1}}.}
If short sales are not allowed then the portfolio is computed numerically using
the function \code{solve.QP()} from the \strong{quadprog} package.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# compute global minimum variance portfolio
gmin_port = globalmin_portfolio(er, covmat)
attributes(gmin_port)
print(gmin_port)
summary(gmin_port, risk_free=r_free)
plot(gmin_port, col="blue")

# compute global minimum variance portfolio with no short sales
gmin_port_ns = globalmin_portfolio(er, covmat, shorts=FALSE)
attributes(gmin_port_ns)
print(gmin_port_ns)
summary(gmin_port_ns, risk_free=r_free)
plot(gmin_port_ns, col="blue")

}
\author{
Eric Zivot
}
