% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jwnDailyPrices}
\alias{jwnDailyPrices}
\title{Daily adjusted closing price for Nordstrom stock}
\format{
A \samp{xts} object on Jan-1993 to Jan-2015 containing:
\describe{
\item{JWN}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=JWN+Historical+Prices}
}
\usage{
jwnDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Nordstrom from Jan-1993 to Jan-2015.
}
\keyword{datasets}
