% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jwnDailyPrices2020}
\alias{jwnDailyPrices2020}
\title{Daily adjusted closing price for Nordstrom stock}
\format{
A \samp{xts} object containing:
\describe{
\item{JWN}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=JWN+Historical+Prices}
}
\usage{
jwnDailyPrices2020
}
\description{
A \samp{xts} object of daily adjusted closing price for Nordstrom from 1993-01-04 to 2020-12-30.
}
\keyword{datasets}
