% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msftMonthlyPrices}
\alias{msftMonthlyPrices}
\title{Monthly adjusted closing price for Microsoft stock}
\format{A \samp{xts} object on Jan-1993 to Dec-2014 containing:
\describe{
  \item{MFST}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=MSFT+Historical+Prices}
}
\usage{
msftMonthlyPrices
}
\description{
A \samp{xts} object of monthly adjusted closing price for Microsoft from Jan-1993 to Dec-2014.
}
\keyword{datasets}

