% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gwn_model.R
\name{plot.gwn_model}
\alias{plot.gwn_model}
\title{Plot method of class gwn_model}
\usage{
\method{plot}{gwn_model}(
  x,
  asset = NULL,
  which.plot = c("residuals", "fitted", "histogram", "acf", "qq", "boxplot",
    "correlation", "risk-return"),
  ...
)
}
\arguments{
\item{x}{object of class gwn_model}

\item{asset}{character, name of asset to plot. If NULL, the first asset is used.}

\item{which.plot}{character, one of "residuals", "fitted", "histogram", "acf",
"qq", "boxplot", "correlation", or "risk-return".}

\item{...}{additional arguments passed to the plotting functions}
}
\value{
A plot of the specified type for the specified asset or for all assets in the model.
}
\description{
Plot method of class gwn_model
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
gwn_fit = estimate_gwn_model(exampleRetS[,"AMZN"])
gwn_fit2 = estimate_gwn_model(exampleRetS)
}
