% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{plot.Markowitz}
\alias{plot.Markowitz}
\title{Plot method of class Markowitz}
\usage{
\method{plot}{Markowitz}(x, plot.assets = FALSE, ...)
}
\arguments{
\item{x}{object of class Markowitz}

\item{plot.assets}{if \samp{TRUE} then plot asset \samp{sd} and \samp{er} with asset name labels}

\item{...}{additional arguments passed to \samp{plot()}}
}
\description{
Plot efficient frontier. The efficient frontier is a plot of portfolio expected return vs.
portfolio standard deviation for a collection of mean-variance efficient portfolios - portfolios
that minimize variance subject to a target expected return.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# tangency portfolio
tan_port <- tangency_portfolio(er, covmat, r_free)
# compute global minimum variance portfolio
gmin_port = globalmin_portfolio(er, covmat)

# compute portfolio frontier
ef <- efficient_frontier(er, covmat, alpha_min=-2,
                         alpha_max=1.5, nport=20)
attributes(ef)

plot(ef)
plot(ef, plot.assets=TRUE, col="blue", pch=16)
points(gmin_port$sd_port, gmin_port$er_port, col="green", pch=16, cex=2)
points(tan_port$sd_port, tan_port$er_port, col="red", pch=16, cex=2)
text(gmin_port$sd_port, gmin_port$er_port, labels="GLOBAL MIN", pos=2)
text(tan_port$sd_port, tan_port$er_port, labels="TANGENCY", pos=2)
sr_tan = (tan_port$er_port - r_free)/tan_port$sd_port
abline(a=r_free, b=sr_tan, col="green", lwd=2)

}
\author{
Eric Zivot
}
