% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gwn_model.R
\name{print.gwn_model}
\alias{print.gwn_model}
\title{Print method for objects of class \code{gwn_model}}
\usage{
\method{print}{gwn_model}(x, ...)
}
\arguments{
\item{x}{object of class \code{gwn_model}}

\item{...}{... additional arguments passed to \samp{print()}}
}
\description{
Print method for objects of class \code{gwn_model}
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
gwn_fit = estimate_gwn_model(exampleRetS[,"AMZN"])
gwn_fit2 = estimate_gwn_model(exampleRetS)
gwn_fit
gwn_fit2
}
