% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{print.Markowitz}
\alias{print.Markowitz}
\title{Print efficient frontier}
\usage{
\method{print}{Markowitz}(x, ...)
}
\arguments{
\item{x}{object of class Markowitz}

\item{...}{additional arguments passed to \samp{print()}}
}
\description{
Print method for \samp{Markowitz} objects.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# tangency portfolio
tan_port <- tangency_portfolio(er, covmat, r_free)
# compute global minimum variance portfolio
gmin_port = globalmin_portfolio(er, covmat)

# compute portfolio frontier
ef <- efficient_frontier(er, covmat, alpha_min=-2,
                         alpha_max=1.5, nport=20)
attributes(ef)
print(ef)

}
\author{
Eric Zivot
}
