% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio.R
\name{print.portfolio}
\alias{print.portfolio}
\title{Print method of class portfolio}
\usage{
\method{print}{portfolio}(x, ...)
}
\arguments{
\item{x}{object of class portfolio}

\item{...}{additional arguments passed to \code{print()}}
}
\description{
Print method for objects of class \code{portfolio}.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# compute equally weighted portfolio
ew = rep(1,3)/3
equal_weight_portfolio = get_portfolio(er=er,cov_mat=covmat,weight_vec=ew)
print(equal_weight_portfolio)

}
\author{
Eric Zivot
}
