% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sr.R
\name{print.SRhat}
\alias{print.SRhat}
\title{Print method of class SRhat}
\usage{
\method{print}{SRhat}(x, ...)
}
\arguments{
\item{x}{object of class SRhat}

\item{...}{additional arguments passed to \code{print()}}
}
\description{
Print method for objects of class \code{SRhat}. Show the estimated
Sharpe ratio, the risk-free rate, and the date range.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
sr_hat = estimate_sr(exampleRetS)
sr_hat

}
\author{
Eric Zivot
}
