% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_var.R
\name{print.VaRhat}
\alias{print.VaRhat}
\title{Print method of class VaRhat}
\usage{
\method{print}{VaRhat}(x, ...)
}
\arguments{
\item{x}{object of class VaRhat}

\item{...}{additional arguments passed to \samp{print()}}
}
\description{
Print method for objects of class \code{VaRhat}. Show the estimated value at risk,
the probability level, the initial investment value, and the method used for
estimating VaR.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
exampleRetC = log(1 + exampleRetS)
var_hat = estimate_var(exampleRetS, alpha=0.05, w=100000,
                      var_method="normal", var_type ="return",
                      return_type="simple")
var_hat

}
\author{
Eric Zivot
}
