% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbuxDailyPrices2020}
\alias{sbuxDailyPrices2020}
\title{Daily adjusted closing price for Starbucks stock}
\format{
A \samp{xts} object containing:
\describe{
\item{SBUX}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=SBUX+Historical+Prices}
}
\usage{
sbuxDailyPrices2020
}
\description{
A \samp{xts} object of daily adjusted closing price for Starbucks from 1993-01-04 to 2020-12-30.
}
\keyword{datasets}
