% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbuxMonthlyPrices}
\alias{sbuxMonthlyPrices}
\title{Monthly adjusted closing price for Starbucks stock}
\format{A \samp{xts} object on Jan-1993 to Dec-2014 containing:
\describe{
  \item{SBUX}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=SBUX+Historical+Prices}
}
\usage{
sbuxMonthlyPrices
}
\description{
A \samp{xts} object of monthly adjusted closing price for Starbucks from Jan-1993 to Dec-2014.
}
\keyword{datasets}

