% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500DailyPrices}
\alias{sp500DailyPrices}
\title{Daily adjusted closing price for S&P 500 index}
\format{
A \samp{xts} object containing:
\describe{
\item{SP500}{adjusted closing price in USD}
}
}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=\%5EGSPC+Historical+Prices}
}
\usage{
sp500DailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for S&P 500 from 1993-01-04 to 2014-12-31.
}
\keyword{datasets}
