% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500MonthlyPrices}
\alias{sp500MonthlyPrices}
\title{Monthly adjusted closing price for S&P 500}
\format{A \samp{xts} object on Jan-1993 to Dec-2014 containing:
\describe{
  \item{SP500}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=\%5EGSPC+Historical+Prices}
}
\usage{
sp500MonthlyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for S&P 500 from Jan-1993 to Dec-2014.
}
\keyword{datasets}

