% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gwn_model.R
\name{summary.gwn_model}
\alias{summary.gwn_model}
\title{Summary method of class gwn_model}
\usage{
\method{summary}{gwn_model}(object, show_correlations = FALSE, ...)
}
\arguments{
\item{object, }{object of class gwn_fit}

\item{show_correlations}{logical, whether to show pairwise correlations}

\item{...}{additional arguments passed to \samp{summary()}}
}
\description{
Show estimated mean and volatilies together with estimated standard errors
and 95\% confidence intervals.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
gwn_fit = estimate_gwn_model(exampleRetS[,"AMZN"])
gwn_fit2 = estimate_gwn_model(exampleRetS)
summary(gwn_fit)
summary(gwn_fit2)

}
\references{
See Zivot, E. (2026) \emph{Introduction to Computational Finance and Financial Econometrics with R},
chapter 7 for a discussion of the Gaussian White Noise model and its estimation.
}
\author{
Eric Zivot
}
