% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{summary.Markowitz}
\alias{summary.Markowitz}
\title{Summary method of class Markowitz}
\usage{
\method{summary}{Markowitz}(object, risk_free = NULL, ...)
}
\arguments{
\item{object}{object of class Markowitz}

\item{risk_free}{numeric, risk free rate}

\item{...}{additional arguments passed to \samp{summary()}}
}
\description{
Summary method for objects of class \samp{Markowitz}. For all portfolios on the efficient
frontier, the expected return, standard deviation and asset weights are shown. If
\samp{risk_free} is given then efficient portfolios that are combinations of the risk free asset
and the tangency portfolio are computed. The class \samp{summary_Markozitz} will be created.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# tangency portfolio
tan_port <- tangency_portfolio(er, covmat, r_free)
# compute global minimum variance portfolio
gmin_port = globalmin_portfolio(er, covmat)

# compute portfolio frontier
ef <- efficient_frontier(er, covmat, alpha_min=-2,
                         alpha_max=1.5, nport=20)
attributes(ef)
summary(ef)

}
\author{
Eric Zivot
}
