% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio.R
\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summary method of class portfolio}
\usage{
\method{summary}{portfolio}(object, risk_free = NULL, ...)
}
\arguments{
\item{object}{object of class portfolio}

\item{risk_free}{numeric, risk free rate}

\item{...}{additional arguments passed to \code{summary()}}
}
\description{
Summary method for objects of class \code{portfolio}. The output is the same as the \code{print}.
If \code{risk_free} is specified then the portfolio Sharpe ratio is also returned.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# compute equally weighted portfolio
ew = rep(1,3)/3
equal_weight_portfolio = get_portfolio(er=er,cov_mat=covmat,weight_vec=ew)
summary(equal_weight_portfolio)

}
\author{
Eric Zivot
}
