% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sr.R
\name{summary.SRhat}
\alias{summary.SRhat}
\title{Summary method of class SRhat}
\usage{
\method{summary}{SRhat}(object, ...)
}
\arguments{
\item{object, }{object of class SRhat}

\item{...}{additional arguments passed to \samp{summary()}}
}
\description{
Show estimated Sharpe ratio with estimated standard errors
and 95\% confidence intervals.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
exampleRetC = log(1 + exampleRetS)
sr_hat = estimate_sr(exampleRetS)
summary(sr_hat)
}
\author{
Eric Zivot
}
