% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_var.R
\name{summary.VaRhat}
\alias{summary.VaRhat}
\title{Summary method of class VaRhat}
\usage{
\method{summary}{VaRhat}(object, ...)
}
\arguments{
\item{object, }{object of class VaRhat}

\item{...}{additional arguments passed to \samp{summary()}}
}
\description{
Show estimated VaR together with estimated standard errors
and 95\% confidence intervals.
}
\examples{
examplePrices = xts::merge.xts(amznDailyPrices, baDailyPrices, costDailyPrices)
examplePrices = xts::to.monthly(examplePrices,OHLC=FALSE)
examplePrices=examplePrices["1998::2002"]
exampleRetS = PerformanceAnalytics::Return.calculate(examplePrices,
                                    method="discrete")
exampleRetS = na.omit(exampleRetS)
exampleRetC = log(1 + exampleRetS)
var_hat = estimate_var(exampleRetS, alpha=0.05, w=100000,
                      var_method="normal", var_type ="return",
                      return_type="simple")
summary(var_hat)
}
\author{
Eric Zivot
}
