% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangency_portfolio.R
\name{tangency_portfolio}
\alias{tangency_portfolio}
\title{Compute tangency portfolio}
\usage{
tangency_portfolio(er_vec, cov_mat, risk_free, shorts = TRUE)
}
\arguments{
\item{er_vec}{\eqn{N \times 1} vector of expected returns \eqn{\boldsymbol{\mu}}}

\item{cov_mat}{\eqn{N \times N} return covariance matrix \eqn{\boldsymbol{\Sigma}}}

\item{risk_free}{numeric, risk free rate \eqn{r_f} used to compute excess returns}

\item{shorts}{logical, if \code{TRUE} then short sales (negative portfolio weights)
are allowed. If \code{FALSE} then no asset is allowed to be sold short.}
}
\value{
A \code{portfolio} object with the following components:
\itemize{
\item \code{call} captures function call
\item \code{er_port} portfolio expected return \eqn{\mu_t \ \mathbf{t}'\boldsymbol{\mu}}
\item \code{sd_port} portfolio standard deviation \eqn{\sigma_t = \sqrt{\mathbf{t}'\boldsymbol{\Sigma} \mathbf{t}}}
\item \code{weight_vec} \eqn{N \times 1} vector of tangency portfolio weights \eqn{\mathbf{t}}
}
}
\description{
Compute tangency (maximum Sharpe ratio) portfolio. The portfolio can allow all assets to be
shorted or not allow any assets to be shorted.
}
\details{
The tangency portfolio \eqn{\mathbf{t}} is the portfolio of risky assets with the highest Sharpe's slope
and solves the optimization problem:
\deqn{ \max_{\mathbf{t}} \frac{\mathbf{t}'\boldsymbol{\mu}-r_f}{(\mathbf{t}'\boldsymbol{\Sigma} \mathbf{t})^{1/2}}}
s.t. \eqn{\mathbf{t}'\mathbf{1}=1,} where \eqn{r_f} denotes the risk-free rate.
If short sales are allowed then there is an analytic solution using matrix algebra of the form
\deqn{\mathbf{t} = \frac{\boldsymbol{\Sigma}^{-1}(\boldsymbol{\mu}-r_f)}{\mathbf{1}'\boldsymbol{\Sigma}^{-1}(\boldsymbol{\mu}-r_f))}.}
If short sales are not allowed then the maximum Sharpe ratio portfolio must be computed numerically
using quadratic programming.
}
\examples{
# construct the data
asset_names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset_names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r_free = 0.005
dimnames(covmat) = list(asset_names, asset_names)

# compute tangency portfolio
tan_port <- tangency_portfolio(er, covmat, r_free)
tan_port
summary(tan_port, risk_free=r_free)
plot(tan_port, col="blue")

# compute tangency portfolio with no short sales
tan_port_ns <- tangency_portfolio(er, covmat, r_free, shorts=FALSE)
tan_port_ns
summary(tan_port_ns, risk_free=r_free)
plot(tan_port_ns, col="blue")

}
\author{
Eric Zivot
}
