% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VanguardPrices}
\alias{VanguardPrices}
\title{Monthly adjusted closing price for Vanguard Mutual Fund}
\format{
A \samp{xts} object containing:
\describe{
\item{vfinx}{Vanguard 500 Index adjusted closing price in USD}
\item{veurx}{Vanguard European Stock Index adjusted closing price in USD}
\item{veiex}{Vanguard Emerging Markets Stock adjusted closing price in USD}
\item{vbltx}{Vanguard Long-Term Bond Index adjusted closing price in USD}
\item{vbisx}{Vanguard Short-Term Bond Index adjusted closing price in USD}
\item{vpacx}{Vanguard Pacific Stock Index adjusted closing price in USD}
}
}
\usage{
VanguardPrices
}
\description{
A \samp{xts} object of monthly adjusted closing price for Vanguard Mutual Fund from Jan-1995 to
Dec-2014.
}
\keyword{datasets}
