% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amznDailyPrices}
\alias{amznDailyPrices}
\title{Daily adjusted closing price for Amazon stock}
\format{A \samp{xts} object on Jan-1998 to Jan-2015 containing:
\describe{
  \item{AMZN}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=AMZN+Historical+Prices}
}
\usage{
amznDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Amazon from Jan-1998 to Jan-2015.
}
\keyword{datasets}

