% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baDailyPrices}
\alias{baDailyPrices}
\title{Daily adjusted closing price for Boeing stock}
\format{A \samp{xts} object on Jan-1993 to Jan-2015 containing:
\describe{
  \item{BA}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=BA+Historical+Prices}
}
\usage{
baDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Boeing from Jan-1993 to Jan-2015.
}
\keyword{datasets}

