% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{costDailyPrices}
\alias{costDailyPrices}
\title{Daily adjusted closing price for Costco stock}
\format{A \samp{xts} object on Jan-1993 to Jan-2015 containing:
\describe{
  \item{COST}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=COST+Historical+Prices}
}
\usage{
costDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Costco from Jan-1993 to Jan-2015.
}
\keyword{datasets}

