% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{costDailyPrices2020}
\alias{costDailyPrices2020}
\title{Daily adjusted closing price for Costco stock}
\format{A \samp{xts} object on Jan-1993 to Jan-2020 containing:
\describe{
  \item{COST}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=COST+Historical+Prices}
}
\usage{
costDailyPrices2020
}
\description{
A \samp{xts} object of daily adjusted closing price for Costco from Jan-1993 to Jan-2020.
}
\keyword{datasets}

