% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/globalMin.portfolio.R
\name{globalMin.portfolio}
\alias{globalMin.portfolio}
\title{Compute global minimum variance portfolio}
\usage{
globalMin.portfolio(er, cov.mat, shorts = TRUE)
}
\arguments{
\item{er}{\samp{N x 1} vector of expected returns}

\item{cov.mat}{\samp{N x N} return covariance matrix}

\item{shorts}{logical, if \samp{TRUE} then short sales (negative portfolio weights)
are allowed. If \samp{FALSE} then no asset is allowed to be sold short.}
}
\value{
\item{call}{captures function call}
 \item{er}{portfolio expected return}
 \item{sd}{portfolio standard deviation}
 \item{weights}{\samp{N x 1} vector of portfolio weights}
}
\description{
Compute global minimum variance portfolio given expected return vector and covariance matrix. The
portfolio can allow all assets to be shorted or not allow any assets to be shorted. The returned
object is of class \samp{portfolio}.
}
\details{
The global minimum variance portfolio \eqn{m} allowing for short sales solves the optimization
problem: min \eqn{t(m)\Sigma m} s.t. \eqn{t(m)1=1} for which there is an analytic solution using
matrix algebra. If short sales are not allowed then the portfolio is computed numerically using
the function \samp{solve.QP()} from the \samp{quadprog} package.
}
\examples{
# construct the data
asset.names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset.names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset.names, asset.names)

# compute global minimum variance portfolio
gmin.port = globalMin.portfolio(er, covmat)
attributes(gmin.port)
print(gmin.port)
summary(gmin.port, risk.free=r.free)
plot(gmin.port, col="blue")

# compute global minimum variance portfolio with no short sales
gmin.port.ns = globalMin.portfolio(er, covmat, shorts=FALSE)
attributes(gmin.port.ns)
print(gmin.port.ns)
summary(gmin.port.ns, risk.free=r.free)
plot(gmin.port.ns, col="blue")
}
\author{
Eric Zivot
}

