% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msftDailyPrices}
\alias{msftDailyPrices}
\title{Daily adjusted closing price for Microsoft stock}
\format{A \samp{xts} object on 1993-01-04 to 2014-12-31 containing:
\describe{
  \item{MFST}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=MSFT+Historical+Prices}
}
\usage{
msftDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Microsoft from 1993-01-04 to 2014-12-31.
}
\keyword{datasets}

