% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.portfolio.R
\name{print.portfolio}
\alias{print.portfolio}
\title{Print method of class portfolio}
\usage{
\method{print}{portfolio}(x, ...)
}
\arguments{
\item{x}{object of class portfolio}

\item{...}{additional arguments passed to \samp{print()}}
}
\description{
Print method for objects of class \samp{portfolio}.
}
\examples{
# construct the data
asset.names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset.names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset.names, asset.names)

# compute equally weighted portfolio
ew = rep(1,3)/3
equalWeight.portfolio = getPortfolio(er=er,cov.mat=covmat,weights=ew)
print(equalWeight.portfolio)
}
\author{
Eric Zivot
}

