% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbuxDailyPrices}
\alias{sbuxDailyPrices}
\title{Daily adjusted closing price for Starbucks}
\format{A \samp{xts} object on 1993-01-04 to 2014-12-31 containing:
\describe{
  \item{SBUX}{adjusted closing price in USD}
}
Indexed by objects of class: [Date] TZ: UTC.}
\source{
Yahoo! Finance: \url{http://finance.yahoo.com/q/hp?s=SBUX+Historical+Prices}
}
\usage{
sbuxDailyPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Starbucks from 1993-01-04 to 2014-12-31.
}
\keyword{datasets}

