% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.portfolio.R
\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summary method of class portfolio}
\usage{
\method{summary}{portfolio}(object, risk.free = NULL, ...)
}
\arguments{
\item{object}{object of class portfolio}

\item{risk.free}{numeric, risk free rate}

\item{...}{controlled variables for \code{summary()}}
}
\description{
Summary method of class portfolio.
}
\examples{
asset.names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset.names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset.names, asset.names)

# compute equally weighted portfolio
ew = rep(1,3)/3
equalWeight.portfolio = getPortfolio(er=er,cov.mat=covmat,weights=ew)
summary(equalWeight.portfolio)
}
\author{
Eric Zivot
}

