% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tangency.portfolio.R
\name{tangency.portfolio}
\alias{tangency.portfolio}
\title{Compute tangency portfolio}
\usage{
tangency.portfolio(er, cov.mat, risk.free, shorts = TRUE)
}
\arguments{
\item{er}{\samp{N x 1} vector of expected returns}

\item{cov.mat}{\samp{N x N} return covariance matrix}

\item{risk.free}{numeric, risk free rate}

\item{shorts}{logical, if \samp{TRUE} then short sales (negative portfolio weights)
are allowed. If \samp{FALSE} then no asset is allowed to be sold short.}
}
\value{
\item{call}{captures function call}
 \item{er}{portfolio expected return}
 \item{sd}{portfolio standard deviation}
 \item{weights}{\samp{N x 1} vector of portfolio weights}
}
\description{
Compute tangency (maximum Sharpe ratio) portfolio. The portfolio can allow all assets to be
shorted or not allow any assets to be shorted.
}
\details{
The tangency portfolio \samp{t} is the portfolio of risky assets with the highest Sharpe's slope
and solves the optimization problem: max \eqn{(t(t)\mu-r_f)/(t(t)\Sigma t^{1/2})} s.t.
\eqn{t(t)1=1} where \eqn{r_f} denotes the risk-free rate. If short sales are allowed then there
is an analytic solution using matrix algebra. If short sales are not allowed then the maximum
Sharpe ratio portfolio must be computed numerically.
}
\examples{
# construct the data
asset.names = c("MSFT", "NORD", "SBUX")
er = c(0.0427, 0.0015, 0.0285)
names(er) = asset.names
covmat = matrix(c(0.0100, 0.0018, 0.0011,
                  0.0018, 0.0109, 0.0026,
                  0.0011, 0.0026, 0.0199),
                nrow=3, ncol=3)
r.free = 0.005
dimnames(covmat) = list(asset.names, asset.names)

# compute tangency portfolio
tan.port <- tangency.portfolio(er, covmat, r.free)
tan.port
summary(tan.port, risk.free=r.free)
plot(tan.port, col="blue")

# compute tangency portfolio with no short sales
tan.port.ns <- tangency.portfolio(er, covmat, r.free, shorts=FALSE)
tan.port.ns
summary(tan.port.ns, risk.free=r.free)
plot(tan.port.ns, col="blue")
}
\author{
Eric Zivot
}

