% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VanguardPrices}
\alias{VanguardPrices}
\title{Monthly adjusted closing price for Vanguard Mutual Fund}
\format{A \samp{xts} object on Jan-1995 to Jan-2015 containing:
\describe{
  \item{vfinx}{Vanguard 500 Index adjusted closing price in USD}
  \item{veurx}{Vanguard European Stock Index adjusted closing price in USD}
  \item{veiex}{Vanguard Emerging Markets Stock adjusted closing price in USD}
  \item{vbltx}{Vanguard Long-Term Bond Index adjusted closing price in USD}
  \item{vbisx}{Vanguard Short-Term Bond Index adjusted closing price in USD}
  \item{vpacx}{Vanguard Pacific Stock Index adjusted closing price in USD}
}
Indexed by class 'yearmon'  num [1:240] 1995 1995 1995 1995 1995 ...}
\usage{
VanguardPrices
}
\description{
A \samp{xts} object of daily adjusted closing price for Vanguard Mutual Fund from Jan-1995 to
Jan-2015.
}
\keyword{datasets}

